Revert "fix(graphs): use weights_only=True when loading graphs in anemoi-grap…"
This reverts commit ab292a52ab938da491bed7fe462b70de6f332802.
