+++ graphs/src/anemoi/graphs/export.py
@@ -17,7 +17,6 @@ from omegaconf import DictConfig
 from torch_geometric.data import HeteroData
 
 from anemoi.graphs.create import GraphCreator
-from anemoi.graphs.utils import load_graph_from_file
 from anemoi.utils.config import DotDict
 
 
@@ -37,7 +36,7 @@ class GraphExporter:
         elif isinstance(graph, (Path, str)):
             graph_path = Path(graph)
             if graph_path.suffix == ".pt":
-                self.graph = load_graph_from_file(graph_path)
+                self.graph = torch.load(graph_path, weights_only=False, map_location="cpu")
             elif graph_path.suffix in (".yaml", ".yml"):
                 self.graph = GraphCreator(graph).create(save_path=None).to("cpu")
             else:
